+++ src/Application/Cli.php
@@ -34,8 +34,8 @@ final class Cli implements Implementation
      * @psalm-mutation-free
      *
      * @param \Closure(OperatingSystem, Environment): Builder $container
-     * @param Sequence<callable(Container, OperatingSystem, Environment): Command> $commands
-     * @param \Closure(Command, Container, OperatingSystem, Environment): Command $mapCommand
+     * @param Sequence<callable(Container): Command> $commands
+     * @param \Closure(Command, Container): Command $mapCommand
      */
     private function __construct(
         private OperatingSystem $os,
@@ -143,13 +143,9 @@ final class Cli implements Implementation
             static fn(
                 Command $command,
                 Container $service,
-                OperatingSystem $os,
-                Environment $env,
             ) => $map(
-                $previous($command, $service, $os, $env),
+                $previous($command, $service),
                 $service,
-                $os,
-                $env,
             ),
         );
     }
@@ -200,14 +196,10 @@ final class Cli implements Implementation
         $mapCommand = static fn(Command $command): Command => $mapCommand(
             $command,
             $container,
-            $os,
-            $env,
         );
         $commands = $this->commands->map(static fn($command) => new Defer(
             \Closure::fromCallable($command),
             $container,
-            $os,
-            $env,
             $mapCommand,
         ));