+++ src/Application/Http.php
@@ -38,9 +38,9 @@ final class Http implements Implementation
      * @psalm-mutation-free
      *
      * @param \Closure(OperatingSystem, Environment): Builder $container
-     * @param Sequence<callable(Pipe, Container, OperatingSystem, Environment): Component<SideEffect, Response>> $routes
+     * @param Sequence<callable(Pipe, Container): Component<SideEffect, Response>> $routes
      * @param \Closure(Component<SideEffect, Response>, Container): Component<SideEffect, Response> $mapRoute
-     * @param Maybe<callable(ServerRequest, Container, OperatingSystem, Environment): Attempt<Response>> $notFound
+     * @param Maybe<callable(ServerRequest, Container): Attempt<Response>> $notFound
      * @param \Closure(ServerRequest, \Throwable, Container): Attempt<Response> $recover
      */
     private function __construct(
@@ -59,7 +59,7 @@ final class Http implements Implementation
      */
     public static function of(OperatingSystem $os, Environment $env): self
     {
-        /** @var Maybe<callable(ServerRequest, Container, OperatingSystem, Environment): Attempt<Response>> */
+        /** @var Maybe<callable(ServerRequest, Container): Attempt<Response>> */
         $notFound = Maybe::nothing();
 
         return new self(
@@ -230,14 +230,12 @@ final class Http implements Implementation
     public function run($input)
     {
         $container = ($this->container)($this->os, $this->env)->build();
-        $os = $this->os;
-        $env = $this->env;
         $mapRoute = $this->mapRoute;
         $recover = $this->recover;
         $pipe = Pipe::new();
         $routes = $this
             ->routes
-            ->map(static fn($handle) => $handle($pipe, $container, $os, $env))
+            ->map(static fn($handle) => $handle($pipe, $container))
             ->map(static fn($component) => $mapRoute($component, $container));
         $router = new Router(
             $routes,
@@ -245,8 +243,6 @@ final class Http implements Implementation
                 static fn($handle) => static fn(ServerRequest $request) => $handle(
                     $request,
                     $container,
-                    $os,
-                    $env,
                 ),
             ),
             static fn($request, $e) => $recover($request, $e, $container),