+++ src/Application/Implementation.php
@@ -63,7 +63,7 @@ interface Implementation
     /**
      * @psalm-mutation-free
      *
-     * @param callable(Container, OperatingSystem, Environment): Command $command
+     * @param callable(Container): Command $command
      *
      * @return self<I, O>
      */
@@ -72,7 +72,7 @@ interface Implementation
     /**
      * @psalm-mutation-free
      *
-     * @param callable(Command, Container, OperatingSystem, Environment): Command $map
+     * @param callable(Command, Container): Command $map
      *
      * @return self<I, O>
      */
@@ -81,7 +81,7 @@ interface Implementation
     /**
      * @psalm-mutation-free
      *
-     * @param callable(Pipe, Container, OperatingSystem, Environment): Component<SideEffect, Response> $handle
+     * @param callable(Pipe, Container): Component<SideEffect, Response> $handle
      *
      * @return self<I, O>
      */
@@ -99,7 +99,7 @@ interface Implementation
     /**
      * @psalm-mutation-free
      *
-     * @param callable(ServerRequest, Container, OperatingSystem, Environment): Attempt<Response> $handle
+     * @param callable(ServerRequest, Container): Attempt<Response> $handle
      *
      * @return self<I, O>
      */