+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - `Innmind\Framework\Application::route()` first parameter must now be expressed via a component inside the callable
 - `Innminf\Framework\Application::notFoundRequestHandler()` callable must now return an `Innmind\Immutable\Attempt<Response>`
 - `Innminf\Framework\Application::notFoundRequestHandler()` has been renamed `::routeNotFound()`
+- `Innmind\Framework\Application::mapCommand()` callable now longer has access to `OperatingSystem` and `Environment` (use services instead)
 
 ### Removed